moving excel functions and
